refactor(cli): extract runtime gate handler
